+++ src/Lifted/Concurrent.hs
@@ -1,8 +1,8 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE Safe              #-}
+{-# LANGUAGE Safe #-}
 
 -- | Concurrency useful and common functions.
-module Concurrent
+
+module Lifted.Concurrent
        ( -- * MVar
          MVar
        , newEmptyMVar
@@ -20,6 +20,7 @@ module Concurrent
        , TVar
        , atomically
        , newTVarIO
+       , readTVarIO
        , STM.modifyTVar'
        , STM.newTVar
        , STM.readTVar
@@ -31,7 +32,7 @@ import qualified Control.Concurrent.MVar     as CCM (newEmptyMVar, newMVar, putM
                                                      readMVar, swapMVar, takeMVar,
                                                      tryPutMVar, tryReadMVar, tryTakeMVar)
 import qualified Control.Concurrent.STM.TVar as STM (modifyTVar', newTVar, newTVarIO,
-                                                     readTVar, writeTVar)
+                                                     readTVar, readTVarIO, writeTVar)
 import qualified Control.Monad.STM           as STM (atomically)
 
 import           Control.Concurrent.MVar     (MVar)
@@ -49,47 +50,47 @@ import           Data.Maybe                  (Maybe)
 -- | Lifted to 'MonadIO' version of 'CCM.newEmptyMVar'.
 newEmptyMVar :: MonadIO m => m (MVar a)
 newEmptyMVar = liftIO CCM.newEmptyMVar
-{-# INLINABLE newEmptyMVar #-}
+{-# INLINE newEmptyMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.newMVar'.
 newMVar :: MonadIO m => a -> m (MVar a)
 newMVar = liftIO . CCM.newMVar
-{-# INLINABLE newMVar #-}
+{-# INLINE newMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.putMVar'.
 putMVar :: MonadIO m => MVar a -> a -> m ()
 putMVar m a = liftIO $ CCM.putMVar m a
-{-# INLINABLE putMVar #-}
+{-# INLINE putMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.readMVar'.
 readMVar :: MonadIO m => MVar a -> m a
 readMVar = liftIO . CCM.readMVar
-{-# INLINABLE readMVar #-}
+{-# INLINE readMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.swapMVar'.
 swapMVar :: MonadIO m => MVar a -> a -> m a
 swapMVar m v = liftIO $ CCM.swapMVar m v
-{-# INLINABLE swapMVar #-}
+{-# INLINE swapMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.takeMVar'.
 takeMVar :: MonadIO m => MVar a -> m a
 takeMVar = liftIO . CCM.takeMVar
-{-# INLINABLE takeMVar #-}
+{-# INLINE takeMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.tryPutMVar'.
 tryPutMVar :: MonadIO m => MVar a -> a -> m Bool
 tryPutMVar m v = liftIO $ CCM.tryPutMVar m v
-{-# INLINABLE tryPutMVar #-}
+{-# INLINE tryPutMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.tryReadMVar'.
 tryReadMVar :: MonadIO m => MVar a -> m (Maybe a)
 tryReadMVar = liftIO . CCM.tryReadMVar
-{-# INLINABLE tryReadMVar #-}
+{-# INLINE tryReadMVar #-}
 
 -- | Lifted to 'MonadIO' version of 'CCM.tryTakeMVar'.
 tryTakeMVar :: MonadIO m => MVar a -> m (Maybe a)
 tryTakeMVar = liftIO . CCM.tryTakeMVar
-{-# INLINABLE tryTakeMVar #-}
+{-# INLINE tryTakeMVar #-}
 
 ----------------------------------------------------------------------------
 -- Lifted STM
@@ -98,9 +99,14 @@ tryTakeMVar = liftIO . CCM.tryTakeMVar
 -- | Lifted to 'MonadIO' version of 'STM.atomically'.
 atomically :: MonadIO m => STM a -> m a
 atomically = liftIO . STM.atomically
-{-# INLINABLE atomically #-}
+{-# INLINE atomically #-}
 
 -- | Lifted to 'MonadIO' version of 'STM.newTVarIO'.
 newTVarIO :: MonadIO m => a -> m (TVar a)
 newTVarIO = liftIO . STM.newTVarIO
-{-# INLINABLE newTVarIO #-}
+{-# INLINE newTVarIO #-}
+
+-- | Lifted to 'MonadIO' version of 'STM.readTVarIO'.
+readTVarIO :: MonadIO m => TVar a -> m a
+readTVarIO = liftIO . STM.readTVarIO
+{-# INLINE readTVarIO #-}