+++ src/Print.hs
@@ -1,11 +1,10 @@
-{-# LANGUAGE ExtendedDefaultRules #-}
-{-# LANGUAGE FlexibleContexts     #-}
-{-# LANGUAGE FlexibleInstances    #-}
-{-# LANGUAGE NoImplicitPrelude    #-}
-{-# LANGUAGE Trustworthy          #-}
-{-# LANGUAGE TypeSynonymInstances #-}
+{-# LANGUAGE FlexibleContexts  #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE Trustworthy       #-}
 
-module Show
+-- | Generalization of 'Prelude.putStr' and 'Prelude.putStrLn' functions.
+
+module Print
        ( Print (..)
        , putText
        , putLText
@@ -13,6 +12,7 @@ module Show
 
 import qualified Base
 import           Data.Function              ((.))
+import qualified Prelude                    as Prelude
 
 import           Control.Monad.IO.Class     (MonadIO, liftIO)
 import qualified Data.ByteString.Char8      as BS
@@ -25,6 +25,7 @@ import qualified Data.Text.Lazy             as TL
 import qualified Data.Text.Lazy.IO          as TL
 
 
+-- | Polymorfic over string and lifted to 'MonadIO' printing functions.
 class Print a where
   putStr :: MonadIO m => a -> m ()
   putStrLn :: MonadIO m => a -> m ()
@@ -46,14 +47,15 @@ instance Print BL.ByteString where
   putStrLn = liftIO . BL.putStrLn
 
 instance Print [Base.Char] where
-  putStr = liftIO . Base.putStr
-  putStrLn = liftIO . Base.putStrLn
+  putStr = liftIO . Prelude.putStr
+  putStrLn = liftIO . Prelude.putStrLn
 
--- For forcing type inference
+-- | Specialized to 'T.Text' version of 'putStrLn' or forcing type inference.
 putText :: MonadIO m => T.Text -> m ()
 putText = putStrLn
 {-# SPECIALIZE putText :: T.Text -> Base.IO () #-}
 
+-- | Specialized to 'TL.Text' version of 'putStrLn' or forcing type inference.
 putLText :: MonadIO m => TL.Text -> m ()
 putLText = putStrLn
 {-# SPECIALIZE putLText :: TL.Text -> Base.IO () #-}