+++ src/TypeOps.hs
@@ -8,6 +8,14 @@
 {-# LANGUAGE TypeFamilies       #-}
 {-# LANGUAGE TypeOperators      #-}
 
+#if __GLASGOW_HASKELL__ <= 710
+{-# LANGUAGE Trustworthy        #-}
+#else
+{-# LANGUAGE Safe               #-}
+#endif
+
+-- | Type operators for writing convenient type signatures.
+
 module TypeOps
        ( type Each
        , type ($)
@@ -26,7 +34,7 @@ import           Control.Type.Operator (type ($), type (<+>))
 -- @
 -- f :: Each [Show, Read] [a, b] => a -> b -> String
 -- =
--- f :: (Show a, Show b, Read a, Show b) => a -> b -> String
+-- f :: (Show a, Show b, Read a, Read b) => a -> b -> String
 -- @
 --
 -- To specify list with single constraint / variable, don't forget to prefix