+++ test_stack_lts.sh
@@ -1,4 +1,5 @@
 set +e
 
-STACK_YAML=stack-7.10.yaml stack build  --no-terminal
-STACK_YAML=stack-8.0.yaml  stack build  --no-terminal
+STACK_YAML=stack-7.10.3.yaml stack build  --no-terminal
+STACK_YAML=stack-8.0.1.yaml  stack build  --no-terminal
+stack build --no-terminal